requirements.txt
ultralytics>=8.3.0
onnxruntime-qnn>=1.23.0
numpy>=1.26.0
opencv-python-headless>=4.5.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
onnx>=1.15.0
